# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/page-sections": {
      "post": {
        "operationId": "PageSectionController_createSection",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "description": "CSRF token from GET /auth/csrf — copy the csrfToken field into this header.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSectionRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created section",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 201
                    },
                    "data": {
                      "$ref": "#/components/schemas/CreateSectionResponseDto"
                    },
                    "message": {
                      "type": "string",
                      "example": "Created"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data",
                    "message"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "message"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Missing permission",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 403
                    },
                    "message": {
                      "type": "string",
                      "example": "Forbidden"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Create a page section",
        "tags": [
          "PageSection"
        ]
      }
    },
    "/admin/page-sections/{id}": {
      "get": {
        "operationId": "PageSectionController_getSection",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page section",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/GetSectionResponseDto"
                    },
                    "message": {
                      "type": "string",
                      "example": "OK"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data",
                    "message"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Missing permission",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 403
                    },
                    "message": {
                      "type": "string",
                      "example": "Forbidden"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Section not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Not Found"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Get a page section with its typed data",
        "tags": [
          "PageSection"
        ]
      }
    },
    "/admin/page-sections/{id}/data": {
      "patch": {
        "operationId": "PageSectionController_updateSectionData",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "description": "CSRF token from GET /auth/csrf — copy the csrfToken field into this header.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePageSectionDataDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated section",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/UpdateSectionDataResponseDto"
                    },
                    "message": {
                      "type": "string",
                      "example": "OK"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data",
                    "message"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid section data or unknown files",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "message"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Missing permission",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 403
                    },
                    "message": {
                      "type": "string",
                      "example": "Forbidden"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Section not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Not Found"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Update the typed data of a page section",
        "tags": [
          "PageSection"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lio API",
    "description": "Authentication and session management API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
...
      "PageSectionType": {
        "type": "string",
        "enum": [
          "SLIDER",
          "PRODUCT_LIST",
          "BANNER",
          "INTRODUCTION",
          "HEADER",
          "FOOTER"
        ]
      },
      "PageSectionLocation": {
        "type": "string",
        "enum": [
          "SLIDER",
          "PRODUCT_LIST",
          "AMAZING_PRODUCTS",
          "BANNER",
          "BANNER_3",
          "BANNER_4",
          "INTRODUCTION",
          "HEADER",
          "FOOTER"
        ],
        "description": "Render location; defaults to the section type when omitted"
      },
      "PageSectionStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "INACTIVE"
        ]
      },
      "CreateSectionRequestDto": {
        "type": "object",
        "properties": {
          "pageId": {
            "type": "number",
            "example": 1,
            "nullable": true,
            "description": "Owning page id; omit for a global (page-less) section"
          },
          "type": {
            "example": "SLIDER",
            "allOf": [
              {
                "$ref": "#/components/schemas/PageSectionType"
              }
            ]
          },
          "location": {
            "example": "SLIDER",
            "description": "Render location; defaults to the section type when omitted",
            "allOf": [
              {
                "$ref": "#/components/schemas/PageSectionLocation"
              }
            ]
          },
          "title": {
            "type": "object",
            "example": "محصولات شگفت‌انگیز",
            "nullable": true
          },
          "link": {
            "type": "object",
            "example": "/products/sale",
            "nullable": true
          },
          "sortOrder": {
            "type": "number",
            "example": 0
          },
          "status": {
            "example": "ACTIVE",
            "allOf": [
              {
                "$ref": "#/components/schemas/PageSectionStatus"
              }
            ]
          }
        },
        "required": [
          "type"
        ]
      },
      "SliderSlideDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 11
          },
          "desktopFileUrl": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "tabletFileUrl": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "mobileFileUrl": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "url": {
            "type": "object",
            "example": null,
            "nullable": true
          }
        },
        "required": [
          "id",
          "desktopFileUrl",
          "tabletFileUrl",
          "mobileFileUrl",
          "url"
        ]
      },
      "SliderSectionDataDto": {
        "type": "object",
        "properties": {
          "slides": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SliderSlideDto"
            }
          }
        },
        "required": [
          "slides"
        ]
      },
...
      "ProductListSectionDataDto": {
        "type": "object",
        "properties": {
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductListItemDto"
            }
          }
        },
        "required": [
          "products"
        ]
      },
...
      "BannerSectionDataDto": {
        "type": "object",
        "properties": {
          "banners": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BannerItemDto"
            }
          }
        },
        "required": [
          "banners"
        ]
      },
      "IntroductionSectionDataDto": {
        "type": "object",
        "properties": {
          "titles": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "example": {
              "title": "عنوان اصلی",
              "subtitle": "توضیح کوتاه"
            }
          },
          "desktopFileUrl": {
            "type": "object",
            "example": "/uploads/images/intro-desktop.png",
            "nullable": true
          },
          "tabletFileUrl": {
            "type": "object",
            "example": "/uploads/images/intro-tablet.png",
            "nullable": true
          },
          "mobileFileUrl": {
            "type": "object",
            "example": "/uploads/images/intro-mobile.png",
            "nullable": true
          }
        },
        "required": [
          "titles",
          "desktopFileUrl",
          "tabletFileUrl",
          "mobileFileUrl"
        ]
      },
...
      "HeaderSectionDataDto": {
        "type": "object",
        "properties": {
          "logo": {
            "$ref": "#/components/schemas/HeaderLogoDto"
          },
          "supportPhone": {
            "type": "object",
            "example": "021-12345678",
            "nullable": true
          },
          "slogan": {
            "type": "object",
            "example": "لیو؛ ساده‌تر خرید کن.",
            "nullable": true
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HeaderItemDto"
            }
          }
        },
        "required": [
          "logo",
          "supportPhone",
          "slogan",
          "items"
        ]
      },
...
      "FooterSectionDataDto": {
        "type": "object",
        "properties": {
          "logo": {
            "$ref": "#/components/schemas/FooterLogoDto"
          },
          "description": {
            "type": "object",
            "example": "فروشگاه اینترنتی لیو؛ تجربهٔ خرید آنلاین سریع، مطمئن و مقرون‌به‌صرفه.",
            "nullable": true
          },
          "slogan": {
            "type": "object",
            "example": "لیو؛ ساده‌تر خرید کن.",
            "nullable": true
          },
          "supportPhone": {
            "type": "object",
            "example": "021-12345678",
            "nullable": true
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FooterLinkDto"
            }
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FooterCategoryDto"
            }
          }
        },
        "required": [
          "logo",
          "description",
          "slogan",
          "supportPhone",
          "links",
          "categories"
        ]
      },
      "CreateSectionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 50
          },
          "pageId": {
            "type": "number",
            "example": 1,
            "nullable": true
          },
          "type": {
            "example": "SLIDER",
            "allOf": [
              {
                "$ref": "#/components/schemas/PageSectionType"
              }
            ]
          },
          "location": {
            "example": "SLIDER",
            "allOf": [
              {
                "$ref": "#/components/schemas/PageSectionLocation"
              }
            ]
          },
          "title": {
            "type": "object",
            "example": "محصولات شگفت‌انگیز",
            "nullable": true
          },
          "link": {
            "type": "object",
            "example": "/products/sale",
            "nullable": true
          },
          "sortOrder": {
            "type": "number",
            "example": 0
          },
          "status": {
            "example": "ACTIVE",
            "allOf": [
              {
                "$ref": "#/components/schemas/PageSectionStatus"
              }
            ]
          },
          "data": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/SliderSectionDataDto"
              },
              {
                "$ref": "#/components/schemas/ProductListSectionDataDto"
              },
              {
                "$ref": "#/components/schemas/BannerSectionDataDto"
              },
              {
                "$ref": "#/components/schemas/IntroductionSectionDataDto"
              },
              {
                "$ref": "#/components/schemas/HeaderSectionDataDto"
              },
              {
                "$ref": "#/components/schemas/FooterSectionDataDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "pageId",
          "type",
          "location",
          "title",
          "link",
          "sortOrder",
          "status",
          "data"
        ]
      },
      "GetSectionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 50
          },
          "pageId": {
            "type": "number",
            "example": 1,
            "nullable": true
          },
          "type": {
            "example": "SLIDER",
            "allOf": [
              {
                "$ref": "#/components/schemas/PageSectionType"
              }
            ]
          },
          "location": {
            "example": "SLIDER",
            "allOf": [
              {
                "$ref": "#/components/schemas/PageSectionLocation"
              }
            ]
          },
          "title": {
            "type": "object",
            "example": "محصولات شگفت‌انگیز",
            "nullable": true
          },
          "link": {
            "type": "object",
            "example": "/products/sale",
            "nullable": true
          },
          "sortOrder": {
            "type": "number",
            "example": 1
          },
          "status": {
            "example": "ACTIVE",
            "allOf": [
              {
                "$ref": "#/components/schemas/PageSectionStatus"
              }
            ]
          },
          "data": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/SliderSectionDataDto"
              },
              {
                "$ref": "#/components/schemas/ProductListSectionDataDto"
              },
              {
                "$ref": "#/components/schemas/BannerSectionDataDto"
              },
              {
                "$ref": "#/components/schemas/IntroductionSectionDataDto"
              },
              {
                "$ref": "#/components/schemas/HeaderSectionDataDto"
              },
              {
                "$ref": "#/components/schemas/FooterSectionDataDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "pageId",
          "type",
          "location",
          "title",
          "link",
          "sortOrder",
          "status",
          "data"
        ]
      },
      "UpdateSliderSlideDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 11
          },
          "desktopFileId": {
            "type": "number",
            "example": 101
          },
          "tabletFileId": {
            "type": "number",
            "example": 102
          },
          "mobileFileId": {
            "type": "number",
            "example": 103
          },
          "url": {
            "type": "object",
            "example": "/products/sale",
            "nullable": true,
            "description": "Relative link target for the slide"
          }
        },
        "required": [
          "id",
          "desktopFileId",
          "tabletFileId",
          "mobileFileId"
        ]
      },
      "UpdateProductListDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 11
          },
          "productId": {
            "type": "number",
            "example": 42
          },
          "sortOrder": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "id",
          "productId"
        ]
      },
      "UpdateBannerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 11
          },
          "title": {
            "type": "string",
            "example": "Summer sale"
          },
          "subtitle": {
            "type": "object",
            "example": "Up to 50% off selected items",
            "nullable": true
          },
          "buttonTitle": {
            "type": "object",
            "example": "Shop now",
            "nullable": true
          },
          "buttonUrl": {
            "type": "object",
            "example": "/products/sale",
            "nullable": true
          },
          "desktopFileId": {
            "type": "number",
            "example": 101
          },
          "tabletFileId": {
            "type": "number",
            "example": 102
          },
          "mobileFileId": {
            "type": "number",
            "example": 103
          }
        },
        "required": [
          "id",
          "title",
          "desktopFileId",
          "tabletFileId",
          "mobileFileId"
        ]
      },
      "UpdateIntroductionDto": {
        "type": "object",
        "properties": {
          "titles": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "example": {
              "title": "عنوان اصلی",
              "subtitle": "توضیح کوتاه"
            }
          },
          "desktopFileId": {
            "type": "number",
            "example": 101
          },
          "tabletFileId": {
            "type": "object",
            "example": 102,
            "nullable": true
          },
          "mobileFileId": {
            "type": "object",
            "example": 103,
            "nullable": true
          }
        },
        "required": [
          "titles",
          "desktopFileId"
        ]
      },
...
      "UpdateHeaderDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "Top-level header items in display order. The list is replaced in full.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdateHeaderItemDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
...
      "UpdateFooterDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "Top-level footer items in display order. The list is replaced in full.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdateFooterItemDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "UpdatePageSectionDataDto": {
        "type": "object",
        "properties": {
          "type": {
            "example": "SLIDER",
            "allOf": [
              {
                "$ref": "#/components/schemas/PageSectionType"
              }
            ]
          },
          "data": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/UpdateSliderSlideDto"
              },
              {
                "$ref": "#/components/schemas/UpdateProductListDto"
              },
              {
                "$ref": "#/components/schemas/UpdateBannerDto"
              },
              {
                "$ref": "#/components/schemas/UpdateIntroductionDto"
              },
              {
                "$ref": "#/components/schemas/UpdateHeaderDto"
              },
              {
                "$ref": "#/components/schemas/UpdateFooterDto"
              }
            ]
          }
        },
        "required": [
          "type",
          "data"
        ]
      },
      "UpdateSectionDataResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 50
          },
          "pageId": {
            "type": "number",
            "example": 1,
            "nullable": true
          },
          "type": {
            "example": "SLIDER",
            "allOf": [
              {
                "$ref": "#/components/schemas/PageSectionType"
              }
            ]
          },
          "location": {
            "example": "SLIDER",
            "allOf": [
              {
                "$ref": "#/components/schemas/PageSectionLocation"
              }
            ]
          },
          "title": {
            "type": "object",
            "example": "محصولات شگفت‌انگیز",
            "nullable": true
          },
          "link": {
            "type": "object",
            "example": "/products/sale",
            "nullable": true
          },
          "sortOrder": {
            "type": "number",
            "example": 1
          },
          "status": {
            "example": "ACTIVE",
            "allOf": [
              {
                "$ref": "#/components/schemas/PageSectionStatus"
              }
            ]
          },
          "data": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/SliderSectionDataDto"
              },
              {
                "$ref": "#/components/schemas/ProductListSectionDataDto"
              },
              {
                "$ref": "#/components/schemas/BannerSectionDataDto"
              },
              {
                "$ref": "#/components/schemas/IntroductionSectionDataDto"
              },
              {
                "$ref": "#/components/schemas/HeaderSectionDataDto"
              },
              {
                "$ref": "#/components/schemas/FooterSectionDataDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "pageId",
          "type",
          "location",
          "title",
          "link",
          "sortOrder",
          "status",
          "data"
        ]
      },
...
    }
  }
}